+++ packages/phishing-controller/src/index.ts
@@ -8,7 +8,8 @@ export type {
 } from './PhishingDetector';
 export { PhishingDetector } from './PhishingDetector';
 export type { PhishingDetectionScanResult, AddressScanResult } from './types';
-export type { TokenScanCacheData, TokenScanResultType } from './types';
+export type { TokenScanCacheData } from './types';
+export { TokenScanResultType } from './types';
 export {
   PhishingDetectorResultType,
   RecommendedAction,